requirements.txt
python-telegram-bot==20.0
requests==2.26.0
tqdm==4.62.3
yt-dlp==2021.12.1
TikTokApi==6.0.0
ffmpeg-python==0.2.0